Retrievable explanations, Restrict explanations to ConceptMaps (#20)
## What is the goal of this PR?
Explanations can exceed the gRPC message size limits. This change allows retrieving explanations separately from the `ConceptMap` answers themselves, allowing us to break the deep nesting and long messages contained in a single answer. 

The paradigm for retriving an explanation is now as follows:
A `ConceptMap` always contains a `pattern` and a boolean flag `hasExplanation`. If the `ConceptMap` has an explanation, a call can be made using message type `Explanation.Req`, returning a list of `ConceptMap` that correspond to to a single layer of the explanation tree.

Unused `Explanation` messages have been removed from answers other than `ConceptMap`.

## What are the changes implemented in this PR?
* `ConceptMap` contains `conceptMap`, `pattern` and `hasExplanation`
* A new RPC endpoint can be performed by clients using type `Explanation.Req`, returning `Explanation.Res`
* `Explanation` removed from answers other than `ConceptMap`