Add prefilling of various response methods (#27)
## What is the goal of this PR?

Currently, the protocol only returns IDs for every answer. However, it is very rare to only require the ID for a concept e.g. for an attribute you are most likely interested in its value. The goal of this PR is to mitigate the case where hundreds of concept IDs are returned for Attributes from a query, and the client program is going to do a sync RPC to get the value of each.

## What are the changes implemented in this PR?

By taking a selection of example queries run against protocol, we can determine that there are some common requests that are always bounded (if the returned result is a concept, pre-filling it wouldn't lead to recursion). In the case that the returned concept would cause this kind of explosion, or for some reason the results can't be pre-fetched, the `isPrefilled` field notifies a client that this concept wasn't pre-filled (as opposed to just missing various pre-filled parts). Currently there is no way to explicitly enable pre-filling.

This is a static and simple variation on the "time travelling RPC" concept, where we are returning the results of another RPC call on the result of a first (see Cap'n Proto), although we are a long way from achieving this as a general pattern right now.