Fixed implementation of Rule and RuleMethod in Concept.proto (#65)
## What is the goal of this PR?

I think you were rushing in implementing `Rule` and `RuleMethod` in `Concept.proto` in PR #64, @flyingsilverfin :) and without enabling CI, you didn't realise that the protocol wasn't building at all.


## What are the changes implemented in this PR?

1) `Rule.SetThen.Res` was not correctly defined: it had an empty `oneof`, which is not a valid syntax.
2) `RuleMethod` was referring to messages that did not exist: `Rule.When` and `Rule.Then` did not exist.
3) `RuleMethod` needed 4 messages each for `req` and `res`: `GetWhen`, `GetThen`, `SetWhen`, and `SetThen` (in that order).
4) `RuleMethod` variable names were not camelCase.

Additionally,
1) `ThingMethod.Req` had a cleanup of message IDs, but some numbers were not properly compacted.

The PR #64 also to reorder the declaration of messages in this file. However, it's partially complete. Notice how under `RuleMethod` there are more message definitions of the subclasses  of `Type`, which presumably was thought be placed at the top of the file with the moving of the `Type` message itself to the very top. I'll fix this in a subsequent PR.