Dissolve the 3 services back into one service named Grakn (#74)
## What is the goal of this PR?

To ensure that services that are inseparable from each other are in fact not separated.

SessionService is useless without a Database from the DatabaseService; TransactionService is useless without a Session from the SessionService. On the server side, the three have a network of dependencies between each other.

## What are the changes implemented in this PR?

Dissolve the 3 services back into one service named Grakn.