Rust gRPC + Protobuf Bindings (#155)
## What is the goal of this PR?

We've added a gRPC + Protobuf compiler for Rust, allowing us to generate Rust bindings for our Protobuf definitions using the `grpc` crate.

## What are the changes implemented in this PR?

We experimented with a couple of different approaches.

First, I tried implementing the compiler using the `tonic` crate. This is the most popular gRPC crate for Rust, and has by far the most recent GitHub activity. However, a limitation of the compiler library I was using appeared to make it impossible to compile multiple `.proto` files at once, which became a blocker. It's possible that that functionality _is_ provided, but to save time, I moved onto the next approach:

Second, I tried implementing the compiler using the `grpc` crate. This is the second most popular gRPC crate for Rust, and has 1.5k GitHub stars (vs `tonic`'s 5k). This one was completed without difficulties, and I've checked it out into the repo in this PR as reference, but we do not intend to use it going forward.

Finally, I found a `rust_grpc_library` rule in `rules_rust`. This is our preferential solution as it delegates all the work of writing the compilation script to an externally-maintained codebase. It doesn't require us to write a single line of Rust code manually, it just works out of the box.

The only limitation is, it is built for the `grpc` crate, our second-favourite crate for the job. We are currently researching whether we can achieve an elegant and working solution using the `tonic` crate.

However, this PR has been open for a while now, and as we have other projects beginning to depend on it - e.g. `factory` - it makes sense to merge these changes upstream and iterate further on it from there.