Add Rust distribution targets and their CI workflows (#168)
## What is the goal of this PR?

We added crate distribution targets (`assemble_crate` and
`deploy_crate`) for Rust, and added CI workflows for deploying snapshot
and release crates.

## What are the changes implemented in this PR?

We recently tried adding an `assemble_crate` rule to our existing Rust
proto codebase to distribute Rust proto bindings, but it didn't work -
it didn't contain any protobuf source code, because the `assemble_crate`
rule doesn't support Cargo build scripts.

But instead of actually adding that support, we had a rethink of the
build architecture. The old idea was (following Tonic's own docs) for
`typedb-protocol` to come bundled with the `.proto` sources and a Cargo
build script that would allow the end-user to compile the protobufs
on-the-fly. This would require them to have `protoc` installed in their
environment, and would dramatically increase their compilation time and
overhead.

A better option would be to bundle the generated Rust sources. So we've
created the `rust_tonic_compile` rule that enables just that in
https://github.com/vaticle/dependencies/pull/391. Now, both
`rust_library` and `assemble_crate` rules can depend on
`typedb_protocol` seamlessly, so we can depend on `typedb_protocol`
either through Bazel or Cargo.

As this PR stabilises the release process, we've gone ahead and added
the relevant CI workflows: `deploy-crate-snapshot` and
`deploy-crate-release`.