Fix crate snapshots failing to deploy due to bad version number (#172)
## What is the goal of this PR?

We fixed a bug where crate snapshots would fail to deploy, due to
passing in an illegal version number.

## What are the changes implemented in this PR?

We've bumped `dependencies` to the latest version to get the fix in:
- https://github.com/vaticle/bazel-distribution/pull/362

It prepends `0.0.0-` to the version number we're deploying, if it looks
like a commit SHA - i.e. when it is a snapshot deployment.

Without this fix, the version number we submit is a commit SHA, which,
by itself, is not a valid crate version, and is rejected by
repo.vaticle.com.