Fix Rust protocol failing to build in Cargo after an IDE sync (#174)
## What is the goal of this PR?

We fixed an issue where the Rust protocol would fail to build in Cargo
after an IDE sync (by bumping `vaticle_dependencies`)

## What are the changes implemented in this PR?

The fix is in:
- https://github.com/vaticle/dependencies/pull/400

Summary: The `typedb-protocol` crate depends heavily on generated
sources (a Rust gRPC + Proto library). The Rust IDE sync tool would
generate these sources if run in the `typedb-protocol` repo itself, but
not if run from a different repo that depended on Protocol. We fixed
that.