Introduce message types to encode annotations (#182)
## What is the goal of this PR?

We replace usages of `boolean` for key arguments with a message type called `Annotation` - this gives us a natural place to extend the set/capabilities of the annotations in the future. The new message type is used to implement the new "unique" annotation.

## What are the changes implemented in this PR?

Introduce `Type.Annotation`, which allows carrying either `Key` or `Unique` annotation. This message type is used through the Concept API protocol specification and replaces all usages of boolean `key` arguments.