Restore and update protocol-nodejs (#188)
## What is the goal of this PR?

We reinstate the `grpc/nodejs` package now that we've reverted from
typedb-driver-nodejs over FFI to a native implementation. The
implementation is updated to use rules_js and rules_ts rather than the
deprecated rules_nodejs, and the protobuf generation is now performed
hermetically.

## What are the changes implemented in this PR?

With protobuf and grpc dependencies upgraded, we also had to update Java
grpc compilation targets. Additionally, we no longer want to deploy a
maven protocol artifact, since the Java driver is now backed by FFI and
TypeDB is not deployed to maven at all.