Unify TypeDB network API & simplify Concept requests (#187)
## What is the goal of this PR?

We merge core and cluster gRPC service definitions in order to unify the
network interface between the two. As part of this change, we simplify
Concept API by merging related requests.

## What are the changes implemented in this PR?

### Build

* migrate to proto3, which allows the use of `optional` in protocol
definitions;
* remove nodejs and python protobuf rules — gRPC API will be provided by
Rust bindings;

### gRPC Services

* merge core and cluster services and structures into one to unify the
API;

### Requests

* merge concept requests where possible by adding arguments (e.g.
`GetOwns`, `GetOwnsExplicit`, and `GetOwnsKey` become `GetOwns` with
`transitivity` and `annotations` as parameters);
* introduce the `Transitivity` enum and the `RolePlayer` message as part
of request restructure.