Add Query::Type to ConceptRowStream (#209)
## Release notes: usage and product changes
In 3.0, we no longer use parsers on our user-facing clients, and we no
longer can differentiate the types of queries we are executing on the
server. However, it can be sometimes useful (e.g. for extra
logging/messaging) to have at least some information about the nature of
the queries.
For these purposes, we add the executed query's type to
`ConceptRowStream`.

## Implementation
For now, it's just general `Read/Write/Schema` (like transaction types),
which is enough for our existing tasks.