Introduce database import and export protocol messages (#224)
## Release notes: usage and product changes
Add `migration` protocol messages for usage in database import and
database export operations:
* a unidirectional `database_export` stream from a TypeDB server to
export a specific database, similar to TypeDB 2.x;
* a bidirectional `databases_import` stream between a client and a
server to import an exported 2.x/3.x TypeDB database into a TypeDB 3.x
server from a client.

The format of migration items used for these operations is an extended
version of TypeDB 2.x's migration items, so it is backward compatible
with 2.x database files. **Important:** it's not intended to import 3.x
databases into 2.x servers.

## Implementation

Add `Migration { Item }` message. The format is an extended version of
the [2.x
protocol](https://github.com/typedb/typedb/blob/2.29.1/migrator/database/data.proto),
so it contains "outdated" fields for its compatibility with old
databases.

Add `Migration { Export }` message. This operation consists of a single
client `Req { database }` and multiple streamed server responses:
 
1. An initial response with the schema.
2. An unlimited number of migration items (multiple messages with
multiple items in one message for potential optimizations)
3. A `Done` message to signal that the server is ready to close the
stream without errors. It can be substituted by a silent stream closure,
but I preferred explicitness here.

Add `Migration { Import }` message. This operation consists of a stream
of client requests:

1. An initial request with the name of the database and its schema
string.
2. An unlimited number of migration items.
3. A `Done` message to signal that the client is finished without
errors, and the server can perform the final validation. This `Done`
message is required and cannot be removed because the client has to
check whether there were finalization errors or not.

and a stream of server responses (actually, there is either a single
`Done` or a single error, but the stream is needed to return errors at
any stage of the communication).