Return the single target server instead of all servers in the initial connection (#242)
## Release notes: usage and product changes
Substitute `ServerManager.All.Res` with `ServerManager.Get.Res` in the
result of the connection opening message. Returning all servers in a
cloud environment does not make sense, since you cannot know the actual
address of other servers until you connect to them (they are probably
hidden behind proxies, and the GRPC protocol does not even know about
them).

So, it makes more sense to return only the target server's server
information, which helps map the connection address with the id of the
replica and all the other parameters for the future requests.

## Implementation